% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_assigning.R
\name{date_sequence}
\alias{date_sequence}
\title{Generates the sequence of dates between specified columns}
\usage{
date_sequence(dataframe, start, end, sequence_name = "date_sequence")
}
\arguments{
\item{dataframe}{A data frame containing date columns used to find the dates between}

\item{start}{The column containing the start date}

\item{end}{The column containing the end date}

\item{sequence_name}{The name of the output column containing the sequence of dates}
}
\value{
A data frame with a columns containing the dates between the specified columns
}
\description{
Generates the sequence of dates between specified columns
}
