% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_assigning.R
\name{find_factored_measurements}
\alias{find_factored_measurements}
\title{Finds variable values associated with a list of dates from an environmental data set}
\usage{
find_factored_measurements(
  data,
  environmental_data,
  date_sequence_column = "date_sequence",
  environment_date_column,
  environment_variable_column,
  variable_column = "variable_values",
  data_fct_col = NA,
  env_fct_col = NA
)
}
\arguments{
\item{data}{A data frame containing a columns with date sequences to be assigned such as those generated by the date_sequence() function}

\item{environmental_data}{A data frame containing the corresponding environmental data to assign}

\item{date_sequence_column}{The column in data which contains the sequence of dates for each row}

\item{environment_date_column}{The column in the environmental data frame which contains the dates the variables were recorded on}

\item{environment_variable_column}{The column of the environmental data frame which contains the variable of interest to assign}

\item{variable_column}{The output column to assign the assigned environmental data to}

\item{data_fct_col}{If the data frame contains information from different environments, such as different nest, the column containing the specifying factor}

\item{env_fct_col}{The column specifying the factor specifying the developmental environment if multiple developmental environments are used}
}
\value{
A data frame containing a column with the assigned environmental data in each row
}
\description{
Finds variable values associated with a list of dates from an environmental data set
}
