% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_assigning.R
\name{generate_values}
\alias{generate_values}
\title{Generating values from the assigned data sequence}
\usage{
generate_values(
  dataframe,
  column_name = "variable_values",
  prefix = NA,
  unit = NA
)
}
\arguments{
\item{dataframe}{A data frame containing assigned developmental environment data such as those generated by find_factored_measurements()}

\item{column_name}{The column containing the assigned data}

\item{prefix}{A prefix for the output column name}

\item{unit}{A unit for the output column name}
}
\value{
A data frame containing distribution data of assigned variable
}
\description{
Generating values from the assigned data sequence
}
